feat(progress): expose real continue reading progress

+++ src/main/java/com/soap/soap/application/model/ContinueReadingItem.java
@@ -3,7 +3,6 @@ package com.soap.soap.application.model;
 import com.soap.soap.domain.model.EditorialLevel;
 import com.soap.soap.domain.model.ReadingOrigin;
 import com.soap.soap.domain.model.ReadingProgressStatus;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -17,7 +16,7 @@ public record ContinueReadingItem(
     String category,
     LocalDateTime startedAt,
     String shortDescription,
-    BigDecimal progressPercentage) {
+    Integer progressPercentage) {
   public ContinueReadingItem(
       UUID readingId,
       String title,
@@ -37,6 +36,6 @@ public record ContinueReadingItem(
         category,
         startedAt,
         null,
-        progressStatus == ReadingProgressStatus.COMPLETED ? BigDecimal.valueOf(100.0) : null);
+        progressStatus == ReadingProgressStatus.COMPLETED ? 100 : null);
   }
 }
